fix(vqf): use single-precision exp in gain calculation
Replace exp with expf in the tau-based gain computation to keep the
implementation aligned with the single-precision math path in vqf.c and
avoid unnecessary double-precision promotion.

+++ src/vqf.c
@@ -270,7 +270,7 @@ static vqf_real_t gainFromTau(vqf_real_t tau, vqf_real_t Ts)
     } else if (tau == (vqf_real_t)(0.0)) {
         return 1; // k=1 for tau=0
     } else {
-        return 1 - exp(-Ts/tau);  // fc = 1/(2*pi*tau)
+        return 1 - expf(-Ts/tau);  // fc = 1/(2*pi*tau)
     }
 }
